+++ CHANGELOG.md
@@ -6,9 +6,11 @@
 * Add delivery statuses - assigned, prepare, enroute, delivered
 * Disable donate and premium links
 * Initial code release 
+* Remove delivered status
 * Remove mass processor
 * Remove Process plugin action link
 * Remove unused settings
+* Rename status labels
 * Set plugin URL to aihr.us
 * Update Aihrus 1.1.0RC1
 * Update plugin url